switch to new structure for esbuild
also builds using esbuild for both desktop and web now instead of tsc for desktop and esbuild to web
