+++ .vscode/launch.json
@@ -16,9 +16,25 @@
 				// "--disable-extensions"
 			],
 			"outFiles": [
-				"${workspaceFolder}/dist/**/*.js"
+				"${workspaceFolder}/out/**/*.js",
+				"!${workspaceFolder}/out/web/**"
 			],
-			"preLaunchTask": "watch"
+			"preLaunchTask": "watch:desktop"
+		},
+		{
+			"name": "Run Web Extension",
+			"type": "extensionHost",
+			"request": "launch",
+			"debugWebWorkerHost": true,
+			"args": [
+				"--extensionDevelopmentPath=${workspaceFolder}",
+				"${workspaceFolder}/playground",
+				"--extensionDevelopmentKind=web"
+			],
+			"outFiles": [
+				"${workspaceFolder}/out/web/**/*.js"
+			],
+			"preLaunchTask": "watch:web"
 		},
 		{
       		"command": "./docs/node_modules/.bin/astro dev",
