+++ .vscode/tasks.json
@@ -17,8 +17,8 @@
 			}
 		},
 		{
-			"label": "watch",
-			"dependsOn": ["npm: watch:tsc", "npm: watch:esbuild"],
+			"label": "watch:desktop",
+			"dependsOn": ["npm: watch:desktop:tsc", "npm: watch:desktop:esbuild"],
 			"presentation": {
 				"reveal": "never"
 			},
@@ -29,11 +29,11 @@
 		},
 		{
 			"type": "npm",
-			"script": "watch:esbuild",
+			"script": "watch:desktop:esbuild",
 			"group": "build",
 			"problemMatcher": "$esbuild-watch",
 			"isBackground": true,
-			"label": "npm: watch:esbuild",
+			"label": "npm: watch:desktop:esbuild",
 			"presentation": {
 				"group": "watch",
 				"reveal": "never"
@@ -41,11 +41,46 @@
 		},
 		{
 			"type": "npm",
-			"script": "watch:tsc",
+			"script": "watch:desktop:tsc",
 			"group": "build",
 			"problemMatcher": "$tsc-watch",
 			"isBackground": true,
-			"label": "npm: watch:tsc",
+			"label": "npm: watch:desktop:tsc",
+			"presentation": {
+				"group": "watch",
+				"reveal": "never"
+			}
+		},
+		{
+			"label": "watch:web",
+			"dependsOn": ["npm: watch:web:tsc", "npm: watch:web:esbuild"],
+			"presentation": {
+				"reveal": "never"
+			},
+			"group": {
+				"kind": "build",
+				"isDefault": true
+			}
+		},
+		{
+			"type": "npm",
+			"script": "watch:web:esbuild",
+			"group": "build",
+			"problemMatcher": "$esbuild-watch",
+			"isBackground": true,
+			"label": "npm: watch:web:esbuild",
+			"presentation": {
+				"group": "watch",
+				"reveal": "never"
+			}
+		},
+		{
+			"type": "npm",
+			"script": "watch:web:tsc",
+			"group": "build",
+			"problemMatcher": "$tsc-watch",
+			"isBackground": true,
+			"label": "npm: watch:web:tsc",
 			"presentation": {
 				"group": "watch",
 				"reveal": "never"
