+++ esbuild-configs/desktop.esbuild.js
@@ -1,0 +1,48 @@
+/* eslint-disable no-undef */
+import { context } from 'esbuild';
+
+export async function desktop(prodFlag, watchFlag) {
+    const ctx = await context({
+        entryPoints: ['src/extension.ts'],
+        bundle: true,
+        format: 'cjs',
+        minify: prodFlag,
+        sourcemap: !prodFlag,
+        sourcesContent: false,
+        platform: 'node',
+        outfile: 'out/extension.js',
+        external: ['vscode'],
+        logLevel: 'warning',
+        plugins: [
+            /* add to the end of plugins array */
+            esbuildProblemMatcherPlugin,
+        ],
+    });
+    if (watchFlag) {
+        await ctx.watch();
+    } else {
+        await ctx.rebuild();
+        await ctx.dispose();
+    }
+}
+
+/**
+ * @type {import('esbuild').Plugin}
+ */
+const esbuildProblemMatcherPlugin = {
+    name: 'esbuild-problem-matcher',
+
+    setup(build) {
+        build.onStart(() => {
+            console.log('[watch] build started');
+        });
+        build.onEnd(result => {
+            result.errors.forEach(({ text, location }) => {
+                console.error(`✘ [ERROR] ${text}`);
+                if (location == null) return;
+                console.error(`    ${location.file}:${location.line}:${location.column}:`);
+            });
+            console.log('[watch] build finished');
+        });
+    },
+};
