+++ esbuild.js
@@ -1,55 +1,49 @@
-import { context } from 'esbuild';
+import { web } from './esbuild-configs/web.esbuild.js';
+import { desktop } from './esbuild-configs/desktop.esbuild.js';
 
 const production = process.argv.includes('--production');
 const watch = process.argv.includes('--watch');
+const modeArgIndex = process.argv.indexOf('--mode');
+const mode = modeArgIndex !== -1 && process.argv[modeArgIndex + 1] ? process.argv[modeArgIndex + 1] : 'default';
 
-async function main() {
-    const ctx = await context({
-        entryPoints: ['src/extension.ts'],
-        bundle: true,
-        format: 'cjs',
-        minify: production,
-        sourcemap: !production,
-        sourcesContent: false,
-        platform: 'node',
-        outfile: 'dist/extension.js',
-        external: ['vscode'],
-        logLevel: 'warning',
-        plugins: [
-            /* add to the end of plugins array */
-            esbuildProblemMatcherPlugin,
-        ],
-    });
-    if (watch) {
-        await ctx.watch();
-    } else {
-        await ctx.rebuild();
-        await ctx.dispose();
+(async () => {
+    try {
+        switch (mode.toLowerCase()) {
+            case 'web':
+                console.log('Running web build...');
+                await web(production, watch).catch(erm => {
+                    console.error(erm);
+                    process.exit(1);
+                });
+                break;
+            case 'desktop':
+                console.log('Running desktop build...');
+                await desktop(production, watch).catch(erm => {
+                    console.error(erm);
+                    process.exit(1);
+                });
+                break;
+            case 'default':
+                // Can't use watch while building both.
+                if (watch) {
+                    console.error('Cannot use flag --watch while building both desktop and web');
+                }
+                await desktop(production, false).catch(erm => {
+                    console.error(erm);
+                    process.exit(1);
+                });
+                await web(production, false).catch(erm => {
+                    console.error(erm);
+                    process.exit(1);
+                });
+                console.log('Build completed.');
+                break;
+            default:
+                console.error(`Unknown mode: ${mode}`);
+                process.exit(1);
+        }
+    } catch (erm) {
+        console.error('Build failed:', erm);
+        process.exit(1);
     }
-}
-
-/**
- * @type {import('esbuild').Plugin}
- */
-const esbuildProblemMatcherPlugin = {
-    name: 'esbuild-problem-matcher',
-
-    setup(build) {
-        build.onStart(() => {
-            console.log('[watch] build started');
-        });
-        build.onEnd(result => {
-            result.errors.forEach(({ text, location }) => {
-                console.error(`✘ [ERROR] ${text}`);
-                if (location == null) return;
-                console.error(`    ${location.file}:${location.line}:${location.column}:`);
-            });
-            console.log('[watch] build finished');
-        });
-    },
-};
-
-main().catch(erm => {
-    console.error(erm);
-    process.exit(1);
-});
+})();
